\name{RidgeVar-package}
\alias{RidgeVar-package}
\alias{RidgeVar}
\docType{package}
\title{
\packageTitle{RidgeVar}
}
\description{
\packageDescription{RidgeVar}
}
\details{
\tabular{ll}{
Package: \tab RidgeVar\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-03-12\cr
License: \tab GPL (>= 2)\cr
}
}
\references{
Dicker, L. H. (2014). Variance estimation in high-dimensional linear models.  Biometrika 101, 269-284.

Dicker, L. H. and Erdogdu, M. A. (2016). Maximum likelihood for variance estimation in high-dimensional linear models. In  Proceedings     of the 19th International Conference on Artificial Intelligence and Statistics (AISTATS 2016), 159-167. JMLR Workshop \eqn{\&} Conference  Proceedings.

Fan, J., Guo, S. and Hao, N. (2012). Variance estimation using refitted cross-validation in ultrahigh-dimensional regression. Journal of Royal Statistical Society Series B 74, 37-65.

Liu, X., Zheng, S. and Feng, X. (2020). Estimation of error variance via ridge regression. Biometrika, 107, 481-488.
}

\keyword{ package }
