\name{RidgeVar-package}
\alias{RidgeVar-package}
\alias{RidgeVar}
\docType{package}
\title{
\packageTitle{RidgeVar}
}
\description{
\packageDescription{RidgeVar}
}
\details{
\tabular{ll}{
Package: \tab RidgeVar\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-03-12\cr
License: \tab GPL (>= 2)\cr
}
}
\references{Liu, X., Zheng, S. and Feng, X. (2019). Estimation of error variance via ridge regression. Manuscript.
}

\keyword{ package }
